+++ runtime/runtime.go
@@ -7,7 +7,9 @@ import (
 	"time"
 
 	"firebase.google.com/go/v4/messaging"
+	"github.com/centrifugal/gocent/v3"
 	valkey "github.com/gomodule/redigo/redis"
+	_ "github.com/lib/pq"
 	"github.com/nyaruka/gocommon/aws/cwatch"
 	"github.com/nyaruka/gocommon/aws/s3x"
 	"github.com/nyaruka/vkutil"
@@ -23,10 +25,11 @@ type Runtime struct {
 	ReadonlyDB *sql.DB
 	VK         *valkey.Pool
 	S3         *s3x.Service
-	ES     *Elastic
-	Dynamo *Dynamo
-	CW     *cwatch.Service
+	ES         *Elastic
+	Dynamo     *Dynamo
+	CW         *cwatch.Service
 	FCM        FCMClient
+	Centrifugo *gocent.Client
 
 	Queues *Queues
 	Stats  *StatsCollector
@@ -84,7 +87,7 @@ func NewRuntime(cfg *Config) (*Runtime, error) {
 	}
 
 	rt.Queues = newQueues(cfg)
-	rt.Stats = NewStatsCollector(rt.VK)
+	rt.Stats = NewStatsCollector(rt.VK, cfg.LatencyExcludedOrgs)
 
 	return rt, nil
 }